Revert "rgw: dump content_range using 64 bit formatters"
This reverts commit cc435e99802f77b3d4b21abe022665ac9df259cf.

Wrong fix; fcgi doesn't do %lld

+++ src/rgw/rgw_rest.h
@@ -189,7 +189,7 @@ extern void dump_content_length(struct req_state *s, size_t len);
 extern void dump_etag(struct req_state *s, const char *etag);
 extern void dump_last_modified(struct req_state *s, time_t t);
 extern void abort_early(struct req_state *s, int err);
-extern void dump_range(struct req_state *s, uint64_t ofs, uint64_t end, uint64_t total_size);
+extern void dump_range(struct req_state *s, off_t ofs, off_t end, size_t total_size);
 extern void dump_continue(struct req_state *s);
 extern void list_all_buckets_end(struct req_state *s);
 extern void dump_time(struct req_state *s, const char *name, time_t *t);
